added permissions into plugins

+++ nanome/api/files.py
@@ -2,6 +2,7 @@ from nanome._internal._files import _Files
 from nanome.util import FileErrorCode
 from nanome._internal._network._commands._callbacks import _Messages
 
+
 class Files(_Files):
     def __init__(self, plugin_instance):
         self.plugin = plugin_instance
@@ -12,6 +13,8 @@ class Files(_Files):
         :param callback: function that will be called with the full filename
         :type callback: method (:class:`~nanome.util.file.FileError`, str) -> None
         """
+        if self.plugin._permissions.get('files', 0 ) <= 0:
+            raise Exception("Plugin requires files permission to use this method.")
         id = self.plugin._network._send(_Messages.print_working_directory, None, callback != None)
         self.plugin._save_callback(id, callback)
 
@@ -23,6 +26,8 @@ class Files(_Files):
         :param callback: called when operation is completed with any potential errors.
         :type callback: method (:class:`~nanome.util.file.FileError`) -> None
         """
+        if self.plugin._permissions.get('files', 0 ) <= 0:
+            raise Exception("Plugin requires files permission to use this method.")
         id = self.plugin._network._send(_Messages.cd, directory, callback != None)
         self.plugin._save_callback(id, callback)
 
@@ -34,6 +39,8 @@ class Files(_Files):
         :param callback: function that will be called with contents of the directory
         :type callback: method (:class:`~nanome.util.file.FileError`, list of :class:`~nanome.util.file.FileMeta`) -> None
         """
+        if self.plugin._permissions.get('files', 0 ) <= 0:
+            raise Exception("Plugin requires files permission to use this method.")
         id = self.plugin._network._send(_Messages.ls, directory, callback != None)
         self.plugin._save_callback(id, callback)
 
@@ -47,6 +54,8 @@ class Files(_Files):
         :param callback: called when operation is completed with any potential errors.
         :type callback: method (:class:`~nanome.util.file.FileError`) -> None
         """
+        if self.plugin._permissions.get('files', 0 ) <= 0:
+            raise Exception("Plugin requires files permission to use this method.")
         id = self.plugin._network._send(_Messages.mv, (source, dest), callback != None)
         self.plugin._save_callback(id, callback)
 
@@ -60,6 +69,9 @@ class Files(_Files):
         :param callback: called with the path to the file and any potential errors
         :type callback: method (:class:`~nanome.util.file.FileError`, str) -> None
         """
+        if self.plugin._permissions.get('files', 0 ) <= 0:
+            raise Exception("Plugin requires files permission to use this method.")
+
         def cb(error, file):
             if (error == FileErrorCode.no_error):
                 with open(dest, 'wb') as ofile:
@@ -79,6 +91,8 @@ class Files(_Files):
         :param callback: called when operation is completed with any potential errors.
         :type callback: method (:class:`~nanome.util.file.FileError`) -> None
         """
+        if self.plugin._permissions.get('files', 0 ) <= 0:
+            raise Exception("Plugin requires files permission to use this method.")
         with open(source, "rb") as f:
             file = f.read()
             f.close()
@@ -93,6 +107,8 @@ class Files(_Files):
         :param callback: called when operation is completed with any potential errors.
         :type callback: method (:class:`~nanome.util.file.FileError`) -> None
         """
+        if self.plugin._permissions.get('files', 0 ) <= 0:
+            raise Exception("Plugin requires files permission to use this method.")
         id = self.plugin._network._send(_Messages.rm, target, callback != None)
         self.plugin._save_callback(id, callback)
 
@@ -104,6 +120,8 @@ class Files(_Files):
         :param callback: called when operation is completed with any potential errors.
         :type callback: method (:class:`~nanome.util.file.FileError`) -> None
         """
+        if self.plugin._permissions.get('files', 0 ) <= 0:
+            raise Exception("Plugin requires files permission to use this method.")
         id = self.plugin._network._send(_Messages.rmdir, target, callback != None)
         self.plugin._save_callback(id, callback)
 
@@ -117,6 +135,8 @@ class Files(_Files):
         :param callback: called when operation is completed with any potential errors.
         :type callback: method (:class:`~nanome.util.file.FileError`) -> None
         """
+        if self.plugin._permissions.get('files', 0 ) <= 0:
+            raise Exception("Plugin requires files permission to use this method.")
         id = self.plugin._network._send(_Messages.cp, (source, dest), callback != None)
         self.plugin._save_callback(id, callback)
 
@@ -128,5 +148,7 @@ class Files(_Files):
         :param callback: called when operation is completed with any potential errors.
         :type callback: method (:class:`~nanome.util.file.FileError`) -> None
         """
+        if self.plugin._permissions.get('files', 0 ) <= 0:
+            raise Exception("Plugin requires files permission to use this method.")
         id = self.plugin._network._send(_Messages.mkdir, target, callback != None)
         self.plugin._save_callback(id, callback)
